requirements.txt
langchain>=0.2.14
langchain-community>=0.2.12
langchain-huggingface>=0.0.3
langchain-openai>=0.1.22
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
flask>=2.3
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
openai>=1.0.0

